[REFACT] Fixed inconsistent types

+++ parser/MappedExe.cpp
@@ -2,7 +2,7 @@
 
 void ExeWrappersContainer::clearWrappers()
 {
-    std::map<int, ExeElementWrapper*>::iterator itr;
+    std::map<size_t, ExeElementWrapper*>::iterator itr;
     for (itr = this->wrappers.begin(); itr != this->wrappers.end(); itr++){
         ExeElementWrapper* wrapper = itr->second;
         delete wrapper;
@@ -10,13 +10,13 @@ void ExeWrappersContainer::clearWrappers()
     wrappers.clear();
 }
 
-ExeElementWrapper* ExeWrappersContainer::getWrapper(int wrapperId)
+ExeElementWrapper* ExeWrappersContainer::getWrapper(size_t wrapperId)
 {
     if (wrappers.find(wrapperId) == wrappers.end()) return NULL;
     return wrappers[wrapperId];
 }
 
-QString ExeWrappersContainer::getWrapperName(int id)
+QString ExeWrappersContainer::getWrapperName(size_t id)
 {
     if (wrappers.find(id) == wrappers.end()) return "";
     return wrappers[id]->getName();
